pyproject.toml
[tool.poetry]
name = "py-django"
version = "0.1.0"
description = ""
authors = ["SaLuTeK1 <romanbabiichuk2811@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django= "^5.0.4"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"



[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
poebuild-backend = "poetry.core.masonry.api"
